Split frontend code into smaller files
