+++ src/common.js
@@ -1,0 +1,9 @@
+const ROOT_ID   = '00000000-0000-0000-0000-000000000000'
+
+function isObject(obj) {
+  return typeof obj === 'object' && obj !== null
+}
+
+module.exports = {
+  ROOT_ID, isObject
+}
