+++ test/frontend_test.js
@@ -1,5 +1,5 @@
 const assert = require('assert')
-const Frontend = require('../src/frontend')
+const Frontend = require('../frontend')
 const ROOT_ID = '00000000-0000-0000-0000-000000000000'
 const uuid = require('../src/uuid')
